+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## UNRELEASE
 
+## 3.3.1
+
+- **Fix** : Fixed Complete payment button on mobile
+- **Fix** : Removed Apple Pay from Android devices
+
 ## 3.3.0
 
 - **Fix** : Fixed currency error on install